Sat_Simulator/condaenv.cvaku5jk.requirements.txt
anyio
argon2-cffi
argon2-cffi-bindings
asttokens
attrs
babel
backcall
beautifulsoup4
bidict
bleach
cffi
charset-normalizer
debugpy
decorator
defusedxml
entrypoints
executing
fastjsonschema
idna
importlib-metadata
importlib-resources
ipykernel
ipython
ipython-genutils
itur
jedi
jinja2
jplephem
json5
jsonschema
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
lxml
markupsafe
matplotlib-inline
mistune
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook-shim
pandocfilters
parso
pexpect
pickleshare
pkgutil-resolve-name
prometheus-client
prompt-toolkit
psutil
ptyprocess
pure-eval
pyastronomy
pycparser
pygments
pyproj
pyrsistent
pyzmq
quantities
requests
send2trash
sgp4
skyfield
sniffio
soupsieve
stack-data
terminado
tinycss2
tomli
traitlets
urllib3
wcwidth
webencodings
websocket-client
zipp